+++ changelog.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2021-06-05
+### Added
+- Task 3. Services and DI
+- ConfigOptionsService injected to the FirstComponent
+- ConstantsService injected to the FirstComponent
+- generatedString injected to the FirstComponent with useFactory GeneratorFactory GeneratorService as a dependency
+- getNewId method added to the GeneratorService
+- LocalStorageService injected to the FirstComponent
+- RandomTextColor directive has been added
+
 ## [0.0.2] - 2021-05-30
 ### Added
 - Task 2. Components